fix: address critical CR feedback on clock/date implementations
Critical fixes from code review:

1. Fix integer overflow in clock.waitsixtieths conversion
   - Cast to int64_t before multiplication to prevent overflow for large values
   - Formula: (ctsixtieths * 1000) / 60 could overflow at ~2.1M sixtieths

2. Remove unsafe unsigned long casts from date arithmetic
   - date.nextweek, date.prevweek, date.tomorrow, date.yesterday
   - Use signed arithmetic (date + 7L*24L*60L*60L) instead of (unsigned long)date + (...)
   - Prevents wrapping issues on 32-bit platforms and handles negative timestamps correctly

3. Add negative value validation to sleep functions
   - clock.sleepfor, clock.waitseconds, clock.waitsixtieths now reject negative durations
   - Error message: "Can't sleep/wait because negative duration is invalid"
   - Follows Frontier convention for error messages

4. Fix misleading DST comment in datetimetoseconds
   - Comment clarified: tm_isdst is ignored by timegm/mkgmtime (always UTC)
   - Previous comment incorrectly suggested DST determination

5. Consolidate epoch offset constant to shared header
   - Define FRONTIER_EPOCH_TO_UNIX_OFFSET = 2082844800LL in timedate.h
   - Replaces local const declarations in multiple functions
   - Single source of truth with documentation: 66 years * 365.25 days/year * 86400 sec/day

All tests pass with no regressions.

+++ tests/headless_date_verbs.c
@@ -273,7 +273,7 @@ static boolean date_valueproc(short token, hdltreenode hparam1,
             if (!getlongvalue(hparam1, 1, &date))
                 return false;
 
-            return setlongvalue((unsigned long)date + (7 * 24 * 60 * 60), vreturned);
+            return setlongvalue(date + (7L * 24L * 60L * 60L), vreturned);
         }
 
         case datv_nextyear: {
@@ -309,7 +309,7 @@ static boolean date_valueproc(short token, hdltreenode hparam1,
             if (!getlongvalue(hparam1, 1, &date))
                 return false;
 
-            return setlongvalue((unsigned long)date - (7 * 24 * 60 * 60), vreturned);
+            return setlongvalue(date - (7L * 24L * 60L * 60L), vreturned);
         }
 
         case datv_prevyear: {
@@ -347,7 +347,7 @@ static boolean date_valueproc(short token, hdltreenode hparam1,
             if (!getlongvalue(hparam1, 1, &date))
                 return false;
 
-            return setlongvalue((unsigned long)date + (24 * 60 * 60), vreturned);
+            return setlongvalue(date + (24L * 60L * 60L), vreturned);
         }
 
         case datv_weeksinmonth: {
@@ -376,7 +376,7 @@ static boolean date_valueproc(short token, hdltreenode hparam1,
             if (!getlongvalue(hparam1, 1, &date))
                 return false;
 
-            return setlongvalue((unsigned long)date - (24 * 60 * 60), vreturned);
+            return setlongvalue(date - (24L * 60L * 60L), vreturned);
         }
 
         case datv_getcurrenttimezone: {
